pyproject.toml
[project]
name = "social-media-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-snowflake>=1.11.1",
    "instagrapi>=2.2.1",
    "instaloader>=4.15.0",
    "loguru>=0.7.3",
    "nltk>=3.9.2",
    "pandas>=2.0.0,<3.0",
    "plotly>=6.5.2",
    "prefect>=3.6.13",
    "proxyproviders>=0.2.1",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.1",
    "snowflake>=1.11.0",
    "snowflake-connector-python>=4.2.0",
    "streamlit>=1.30.0",
    "wordcloud>=1.9.6",
]
